move stateless agg impls into standalone module
