+++ src/stream/src/executor/aggregation/agg_impl/foldable.rs
