+++ src/stream/src/executor/aggregation/agg_impl/row_count.rs
